View Changes 	- Imrpoved XAML for Merges. 	- An Adress can be entered

+++ WeatherApplication/WeatherObjects/Main.cs
@@ -5,7 +5,7 @@ namespace WeatherObjects
     {
         public double temp { get; set; }
         public int humidity { get; set; }
-        public int pressure { get; set; }
+        public double pressure { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
